requirements.txt
fastapi
uvicorn
pandas
numpy
networkx
scipy
pydantic
pyyaml
python-multipart
python-jose[cryptography]